+++ src/Templates/SimpleDDD/content/Backend/Issues/Mappings/IssueMappings.cs
@@ -1,0 +1,13 @@
+﻿using AutoMapper;
+using GoldenEye.Shared.Core.Mappings;
+
+namespace GoldenEye.WebApi.Template.SimpleDDD.Backend.Issues.Mappings
+{
+    internal class IssueMappings : Profile, IMappingDefinition
+    {
+        public IssueMappings()
+        {
+            //Place here your automapper settings
+        }
+    }
+}