+++ src/Templates/SimpleDDD/content/Contracts/Issues/Events/IssueDeleted.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using GoldenEye.Backend.Core.DDD.Events;
+
+namespace GoldenEye.WebApi.Template.SimpleDDD.Contracts.Issues.Events
+{
+    public class IssueDeleted : IEvent
+    {
+        public Guid IssueId { get; }
+
+        public Guid StreamId => IssueId;
+
+        public IssueDeleted(Guid issueId)
+        {
+            IssueId = issueId;
+        }
+    }
+}