+++ src/Templates/SimpleDDD/content/Contracts/Issues/Queries/GetIssues.cs
@@ -1,0 +1,8 @@
+﻿using GoldenEye.Backend.Core.DDD.Queries;
+
+namespace GoldenEye.WebApi.Template.SimpleDDD.Contracts.Issues.Queries
+{
+    public class GetIssues : IListQuery<Views.IssueView>
+    {
+    }
+}